# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TZONA Public API
  version: 2.0.0
  description: |
    Документация описывает Telegram-first API TZONA V2. Все успешные ответы возвращают
    стандартную обёртку `{ "success": true, "data": ... }`, а ошибки используют
    структуру `{ "success": false, "error": { code, message, classification, details } }`.
    Для защищённых эндпоинтов требуется JWT токен в заголовке `Authorization: Bearer <token>`
    и CSRF токен в заголовке `X-Csrf-Token`.
servers:
  - url: https://api.tzona.app
    description: Production
  - url: http://localhost:4000
    description: Local development
  - url: https://analytics.tzona.app/api
    description: Analytics microservice
security:
  - bearerAuth: []
...
paths:
...
  /stats/{userId}:
    servers:
      - url: https://analytics.tzona.app/api
        description: Analytics microservice
    get:
      tags: [Analytics]
      summary: Агрегированная статистика по профилю
      description: Возвращает количество сессий, выполненных упражнений и распределения RPE для заданного профиля.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Статистика профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsStatsResponse'
        '400':
          description: Некорректный идентификатор профиля
        '500':
          description: Ошибка вычислений или БД
  /batch/profile-stats:
    servers:
      - url: https://analytics.tzona.app/api
        description: Analytics microservice
    post:
      tags: [Analytics]
      summary: Батч-расчёт статистики по нескольким профилям
      description: Позволяет запросить статистику сразу для нескольких профилей за один запрос. Сервис автоматически ограничивает объём выборки и переиспользует кэш для уже рассчитанных профилей.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsBatchStatsRequest'
      responses:
        '200':
          description: Батч-ответ со статистикой
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsBatchStatsResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          description: Ошибка вычислений или БД
  /aggregate:
    servers:
      - url: https://analytics.tzona.app/api
        description: Analytics microservice
    get:
      tags: [Analytics]
      summary: Общая статистика платформы
      description: Возвращает ключевые метрики платформы, недельные тренды и ТОП упражнений для аналитических дашбордов.
      responses:
        '200':
          description: Сводка по платформе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsAggregateResponse'
        '500':
          description: Ошибка вычислений или БД
  /trends/profile/{userId}:
    servers:
      - url: https://analytics.tzona.app/api
        description: Analytics microservice
    get:
      tags: [Analytics]
      summary: Недельные тренды и прогнозы по профилю
      description: Возвращает временной ряд и прогноз по объёму, количеству сессий и коэффициенту завершения для заданного профиля.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Прогноз по профилю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileTrendResponse'
        '400':
          description: Некорректный идентификатор профиля
        '500':
          description: Ошибка аналитики или БД
  /trends/platform:
    servers:
      - url: https://analytics.tzona.app/api
        description: Analytics microservice
    get:
      tags: [Analytics]
      summary: Общие тренды платформы
      description: Возвращает недельные метрики платформы и прогноз следующей недели на основе линейного тренда.
      responses:
        '200':
          description: Прогноз платформы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformTrendResponse'
        '500':
          description: Ошибка аналитики или БД
...
  /export:
    servers:
      - url: https://analytics.tzona.app/api
        description: Analytics microservice
    get:
      tags: [Analytics]
      summary: Экспорт аналитики в JSON или CSV
      description: >
        Генерирует экспорт статистики, агрегатов или трендов. Для профилей требуется `profileId`. Поддерживаются форматы JSON
        и CSV (с автоматическим формированием `Content-Disposition`).
      parameters:
        - name: resource
          in: query
          required: true
          schema:
            type: string
            enum: [profile_stats, aggregate, profile_trends, platform_trends]
          description: Тип ресурса для экспорта.
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
          description: Желаемый формат выгрузки.
        - name: profileId
          in: query
          schema:
            type: string
            format: uuid
          description: Укажите UUID профиля для `profile_stats` и `profile_trends`.
      responses:
        '200':
          description: Экспортированные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsExportEnvelope'
            text/csv:
              schema:
                type: string
                format: binary
                description: CSV документ с таблицей экспортированных строк.
        '400':
          description: Некорректные параметры запроса
        '500':
          description: Ошибка аналитики или БД
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorObject:
      type: object
      required: [code, message, status]
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: integer
        classification:
          type: string
          enum: [validation, authentication, authorization, business, dependencies, internal, not_found]
        details:
          type: object
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...
    AnalyticsWeeklyStat:
      type: object
      properties:
        weekStart:
          type: string
          format: date
        completedSessions:
          type: integer
        totalSessions:
          type: integer
        totalVolume:
          type: integer
    AnalyticsProgressPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        totalEntries:
          type: integer
        buckets:
          type: object
          additionalProperties:
            type: integer
    AnalyticsStatsResponse:
      type: object
      properties:
        totalSessions:
          type: integer
        totalExercises:
          type: integer
        averagePerformance:
          type: number
          format: float
        weeklyStats:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsWeeklyStat'
        progressOverTime:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsProgressPoint'
    AnalyticsBatchStatsRequest:
      type: object
      required: [profileIds]
      properties:
        profileIds:
          type: array
          minItems: 1
          maxItems: 25
          description: Список UUID профилей. При превышении лимита сервис обрежет список до разрешённого значения.
          items:
            type: string
            format: uuid
    AnalyticsBatchStatsResponse:
      type: object
      required: [requested, processed, limit, results]
      properties:
        requested:
          type: integer
          minimum: 0
        processed:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          description: Текущий лимит профилей, обрабатываемых за один батч.
        results:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AnalyticsStatsResponse'
    AnalyticsAggregateSummary:
      type: object
      properties:
        totalProfiles:
          type: integer
        activeProfiles7d:
          type: integer
        totalSessions:
          type: integer
        completedSessions:
          type: integer
        totalVolume:
          type: integer
        averageVolumePerSession:
          type: number
          format: float
    AnalyticsAggregateTrend:
      type: object
      properties:
        weekStart:
          type: string
          format: date
        totalSessions:
          type: integer
        completedSessions:
          type: integer
        uniqueProfiles:
          type: integer
        totalVolume:
          type: integer
...
    ProfileTrendResponse:
      type: object
      properties:
        profileId:
          type: string
          format: uuid
        generatedAt:
          type: string
          format: date-time
        windowWeeks:
          type: integer
        weeklySeries:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsWeeklyStat'
        insights:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsTrendInsight'
    PlatformTrendResponse:
      type: object
      properties:
        scope:
          type: string
          example: platform
        generatedAt:
          type: string
          format: date-time
        windowWeeks:
          type: integer
        weeklySeries:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsWeeklyStat'
        insights:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsTrendInsight'
    AnalyticsTopExercise:
      type: object
      properties:
        exerciseKey:
          type: string
        usageCount:
          type: integer
    AnalyticsAggregateResponse:
      type: object
      properties:
        summary:
          $ref: '#/components/schemas/AnalyticsAggregateSummary'
        weeklyTrends:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsAggregateTrend'
        topExercises:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsTopExercise'
    AnalyticsExportEnvelope:
      type: object
      required: [resource, format, generatedAt]
      properties:
        resource:
          type: string
          enum: [profile_stats, aggregate, profile_trends, platform_trends]
        format:
          type: string
          enum: [json, csv]
        generatedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        data:
          description: Полезная нагрузка присутствует только для JSON-ответов
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/AnalyticsStatsResponse'
            - $ref: '#/components/schemas/AnalyticsAggregateResponse'
            - $ref: '#/components/schemas/ProfileTrendResponse'
            - $ref: '#/components/schemas/PlatformTrendResponse'
...